Area of Circle

+++ Exercise_2/2.2.py
@@ -1,3 +1,3 @@
-rad = int(input("Enter radius :"))
+rad = float(input("Enter radius :"))
 area = 3.14*rad**2
 print("area of circle is",area)
